+++ relxg_3.0.2/user/demodu.h
@@ -27,7 +27,7 @@ void process_buffer_and_sum(float *input_buffer, int buffer_size);
 void frm_sych(void);
 void demodu_process(void);
 
-extern int da_index;
+extern uint16_t da_index;
 extern uint8_t da[MAX_COUNT];
 extern uint8_t hex[HEX_SIZE];
 extern uint16_t TIM7_cnt;
